+++ schemas/index.ts
@@ -1,0 +1,24 @@
+/**
+ * 스키마 통합 export 파일
+ */
+
+// Admin 스키마
+export * from './adminSchema';
+
+// Cash Transaction 스키마
+export * from './cashTransactionSchema';
+
+// Category 스키마
+export * from './categorySchema';
+
+// Discount 스키마
+export * from './discountSchema';
+
+// Ledger 스키마
+export * from './ledgerSchema';
+
+// Menu 스키마
+export * from './menuSchema';
+
+// Staff 스키마
+export * from './staffSchema';
